Detect when a job is lost inside the tenacity exception and raise lost inundate jobs to pipeline failure
Some polling errors being caught by broadest exception in polling
function of NomadJobManager. So checked the wrapped exception to see if
it contains URLNotFound and if it does mark as lost.

Also make lost jobs in the pipeline stage a failure for the pipeline and
exit before beginning the other stages to save time.