Add tests verifying metrics aggregation deduplication (#7)
* Add tests verifying metrics aggregation deduplication

Added unit tests and test data to verify that the current
MetricsAggregator class successfully deduplicates metrics for a unique
"collection_id", "stac_item_id", "scenario" combination only in the case
where an exact or near duplicate is present. In the case that a set of
index columns has different metrics an error is raised indicating a
violation of idempotency due to code or data changes.

Also added a test docker compose file for the unit tests and updated the
README with instructions on running the tests. Since the tests passed I
went ahead and deleted the clean_agg_metrics.py script from tools since
its primary job was getting rid of near duplicates.

* Update README.md