Add aoi_stac_item_id and aoi_geom_path arguments
The goal of this commit is to refactor the pipeline,
submit_stac_batch.py, and the pipeline job definitions to not need an
AOI gpkg when --aoi_stac_item_id is provided along with a valid STAC
item ID string. Previously --aoi_is_item was a logical flag that
extracted the STAC item ID from the AOI gpkg name. When this flag was on
then the STAC item was queried by item ID instead of geometry.
--aoi_is_item was changed to --aoi_stac_item_id and was modified to
accept a string and that string will be used to query STAC items by item
ID and then extract an AOI within the pipeline itself from the benchmark
STAC when aoi_stac_item_id is being used.

Pipeline code was changed to extract the geometry from the STAC item
provided in the --aoi_stac_item_id argument. The geometry will be held
in GeoDataFrame and won't be persisted to disk. The --aoi argument will
become optional and was changed to --aoi_geom_path. So instead of
requiring --aoi we now require one of --aoi_geom_path or
--aoi_stac_item_id. When a --aoi_stac_item_id string is provided then
that string will be used when writing the pipeline outputs instead of
pulling the aoi name from the 'aoi_name' tag provided by the user.

submit_stac_batch.py was also changed. The big change to this script was
that we won't be extracting geometries from the benchmark STAC in this
script.

Nomad job definitions were changed to remove the aoi meta and add
optional --aoi_stac_item_id and --aoi_geom_path as optional meta
parameters. Conditional logic was added using a templated wrapper script
defined in the job definition around the command that invokes the
pipeline in the pipeline Nomad job depending on which meta parameter the
user provides.