Add aoi_name tag derivation from aoi arguments
Modified the way the aoi_name tag is handled. The pipeline first looks
to see if it has been provided by a user and if it has that aoi_name tag
gets precedence. If it hasn't been provided then the aoi_name tag is
derived from aoi_stac_item_id or aoi_geom_path (whichever was provided)