Add aoi_stac_item_id and aoi_geom_path arguments (#9)
* Add aoi_stac_item_id and aoi_geom_path arguments

The goal of this commit is to refactor the pipeline,
submit_stac_batch.py, and the pipeline job definitions to not need an
AOI gpkg when --aoi_stac_item_id is provided along with a valid STAC
item ID string. Previously --aoi_is_item was a logical flag that
extracted the STAC item ID from the AOI gpkg name. When this flag was on
then the STAC item was queried by item ID instead of geometry.
--aoi_is_item was changed to --aoi_stac_item_id and was modified to
accept a string and that string will be used to query STAC items by item
ID and then extract an AOI within the pipeline itself from the benchmark
STAC when aoi_stac_item_id is being used.

Pipeline code was changed to extract the geometry from the STAC item
provided in the --aoi_stac_item_id argument. The geometry will be held
in GeoDataFrame and won't be persisted to disk. The --aoi argument will
become optional and was changed to --aoi_geom_path. So instead of
requiring --aoi we now require one of --aoi_geom_path or
--aoi_stac_item_id. When a --aoi_stac_item_id string is provided then
that string will be used when writing the pipeline outputs instead of
pulling the aoi name from the 'aoi_name' tag provided by the user.

submit_stac_batch.py was also changed. The big change to this script was
that we won't be extracting geometries from the benchmark STAC in this
script.

Nomad job definitions were changed to remove the aoi meta and add
optional --aoi_stac_item_id and --aoi_geom_path as optional meta
parameters. Conditional logic was added using a templated wrapper script
defined in the job definition around the command that invokes the
pipeline in the pipeline Nomad job depending on which meta parameter the
user provides.

* Fix conditional aoi argument passing in Nomad job definitions

The previous commit's changes to the nomad job definitions broke the
pipeline job. This commit's changes to the nomad job definitions
successfully allow for conditional dispatch of the coordinator task
depending on which aoi argument has been submitted to the parameterized
job.

I also fixed a small indentation bug in data_service.py and reformatted
according to the repos agreed upon line length conventions

* Update README

Updated README.md in repo root to reflect the fact that for the test
pipeline to fully work it still currently needs access to the fimc-data
bucket so that the agreement job can access masks.

* Remove initiating pipeline args into environment

Removed the NOMAD_META variables related to the pipeline job from the
env stanza. They aren't necessary for the task to call the pipeline's
main.py

* Remove conditional argument handling from pipeline job def

Removed the complexity of the inline bash script from pipeline job
definition. Now all optional meta parameters that are related to the
call to main.py are fed into the coordinator task. When a parameter
isn't provided main.py is fed an empty string for that argument and
resolves the argument to none during pipeline initializiation

* Add aoi_name tag derivation from aoi arguments

Modified the way the aoi_name tag is handled. The pipeline first looks
to see if it has been provided by a user and if it has that aoi_name tag
gets precedence. If it hasn't been provided then the aoi_name tag is
derived from aoi_stac_item_id or aoi_geom_path (whichever was provided)

* Update comment for tags meta parameter